requirements.txt
pandas==2.1.4
python-binance==1.0.19
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
